moved Registry under Support
